core/requirements.txt
# Package         Version
# --------------- ---------
python-jenkins
django-cors-headers == 3.6.0
dj_rest_auth == 2.1.3
djangorestframework == 3.12.2 
certifi == 2020.12.5 
cffi == 1.14.4
djangorestframework-simplejwt == 4.6.0 
chardet == 4.0.0 
cryptography == 3.3.1 
defusedxml == 0.6.0 
django-allauth == 0.44.0 
idna == 2.10 
oauthlib == 3.1.0
pycparser == 2.20 
pyjwt == 2.0.0 
python3-openid == 3.2.0 
requests == 2.25.1 
requests-oauthlib == 1.3.0 
six == 1.15.0 
urllib3 == 1.26.2
stripe
apscheduler
